td-cloud/requirements.txt
flask
flask-socketio
eventlet
gunicorn
redis
eventlet

google-cloud-pubsub
google-cloud-tasks
google-cloud-storage